Change variable type

+++ sbnd/noise_analysis/TPC_coherent_noise.cpp
@@ -137,7 +137,7 @@ void LoadRawDigits(TFile *inFile)
 
 			//If channel is responsive the channel will grab the noise 
 			bool skip_channel = false;
-			vector<double> x(myADC[index].Samples(),0);
+			vector<short> x(myADC[index].Samples(),0);
 			for (size_t itick=0; itick < myADC[index].Samples(); ++itick){ 
 				if (abs(myADC[index].ADC(itick)-myADC[index].GetPedestal()) >  10){
 					skip_channel = true;
